Base model added and Ids converted to long

+++ src/main/java/com/halisaha/backend/repository/PitchRepository.java
@@ -3,5 +3,5 @@ package com.halisaha.backend.repository;
 import com.halisaha.backend.model.Pitch;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PitchRepository extends JpaRepository<Pitch, Integer> {
+public interface PitchRepository extends JpaRepository<Pitch, Long> {
 }
