+++ src/main/java/com/halisaha/backend/repository/ReservationRepository.java
@@ -7,11 +7,11 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
 
-public interface ReservationRepository extends JpaRepository<Reservation, Integer> {
+public interface ReservationRepository extends JpaRepository<Reservation, Long> {
 
     // DİKKAT: Artık PitchId değil, SubPitchId üzerinden kontrol yapıyoruz
     boolean existsBySubPitchIdAndBookingDateAndStartTimeLessThanAndEndTimeGreaterThan(
-            Integer subPitchId,
+            Long subPitchId,
             LocalDate bookingDate,
             LocalTime endTime,
             LocalTime startTime
